# server/swagger-auto-generated.yaml
openapi: 3.0.2
info:
  title: Campus Events and Entertainment Center API
  version: 1.0.0
  description: API for managing campus events and entertainment
servers:
  - url: http://localhost:5000/api
  - url: http://217.142.229.202/api
  - url: http://10.27.41.93:5000/api
security:
  - bearerAuth: []
paths:
...
  /events:
    post:
      tags:
        - Events
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - title
                - description
                - poster
                - publish_organization
                - participants
                - start_time
                - end_time
                - location
                - status
              properties:
                title:
                  type: string
                description:
                  type: string
                poster:
                  type: string
                  description: The URL of the event's poster
                publish_organization:
                  type: string
                  description: The organization that published the event
                participants:
                  type: array
                  items:
                    $ref: '#/components/schemas/EventParticipant'
                start_time:
                  type: string
                  format: date-time
                end_time:
                  type: string
                  format: date-time
                tags:
                  type: array
                  items:
                    $ref: '#/components/schemas/EventTag'
                location:
                  type: string
                capacity:
                  type: integer
      responses:
        '201':
          description: Event created successfully
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
    get:
      tags:
        - Events
      summary: Get a list of all events
      responses:
        '200':
          description: Event list found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '500':
          $ref: '#/components/responses/500'
  /events/{event_id}:
    get:
      tags:
        - Events
      summary: Get an event by id
      parameters:
        - $ref: '#/components/parameters/path_event_id'
      responses:
        '200':
          description: Event found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          $ref: '#/components/responses/404'
    delete:
      tags:
        - Events
      summary: Delete an event by id
      parameters:
        - $ref: '#/components/parameters/path_event_id'
      responses:
        '204':
          description: Event deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          $ref: '#/components/responses/404'
  /events-tags:
    post:
      tags:
        - Events
      summary: Create a new event tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - tag_name
              properties:
                tag_name:
                  type: string
                  description: The name of the tag
                  example: Music
                  minLength: 1
                  maxLength: 50
                  pattern: ^[a-zA-Z0-9_\-\ ]+$
      responses:
        '201':
          description: Event tag created successfully
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    get:
      tags:
        - Events
      summary: Get all event tags
      responses:
        '200':
          description: Event tag list found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '500':
          $ref: '#/components/responses/500'
  /event-reservations/{event_id}:
    post:
      tags:
        - Events
      summary: Reserve a spot for an event
      parameters:
        - $ref: '#/components/parameters/path_event_id'
      responses:
        '201':
          description: Reservation successful
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
...
components:
  parameters:
...
    path_event_id:
      name: event_id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Event:
      type: object
      additionalProperties: false
      required:
        - id
        - title
        - description
        - poster
        - organizer
        - publish_organization
        - participants
        - time
        - location
        - tags
        - status
        - comments
        - ratings
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
        description:
          type: string
        poster:
          type: string
          description: The URL of the event's poster
        organizer_id:
          type: integer
        publish_organization:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        status:
          type: integer
        location:
          type: string
        capacity:
          type: integer
    EventParticipant:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - description
        - avatar
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
        description:
          type: string
        avatar:
          type: string
...
    EventTag:
      type: object
      additionalProperties: false
      required:
        - id
        - name
      properties:
        id:
          type: integer
          description: The auto-generated id of the tag
          readOnly: true
        name:
          type: string
          description: The name of the tag
...
    ErrorDetail:
      title: ErrorDetail
      additionalProperties: false
      required:
        - title
        - description
      type: object
      properties:
        title:
          title: Title
          type: string
        description:
          title: Description
          type: string
...
  responses:
    '400':
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
    '401':
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
...
    '404':
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
    '429':
      description: Too Many Requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
    '500':
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
...